pyproject.toml
[project]
name = "md-student"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "lxml>=5.3.1",
    "markdown>=3.7",
    "openai>=1.69.0",
    "pydantic-settings>=2.8.1",
    "pyyaml>=6.0.2",
    "toga>=0.5.0",
]
